+++ CHANGES.md
@@ -254,6 +254,8 @@
 * Fixed issue where winsimple-controls.hpp was required due to ws::Window::handleNotifyForChildren.
 * Added Steering Behavior Demo
 * Added ws::Texture::setSize() and ws::Texture::setScale for image pixel data resizing.
+* Fixed Custom window class crashing.
+ 
 
 # New major features to be added.